feat: implement ClientHandlerThread for managing client connections and enhance ThreadPool to utilize it
