+++ server/common/thread_pool.py
@@ -1,18 +1,21 @@
-import logging
 import threading
-from common.utils import has_won, load_bets, store_bets
-from mbp.ack import WinnersAck
-from mbp.bet import rcv_bets_in_batch
 
 
 class ThreadPool:
-    def __init__(self, num_threads, file_lock, barrier):
-        self.threads = [ClientHandlerThread(i, file_lock, barrier) for i in range(num_threads)]
+    def __init__(self, threads):
+        self.threads = threads
+        self.lock = threading.Lock()
+        self.by_ip = {}
+        self.next = 0
 
     def assign_connection(self, client_socket, ip):
-        thread_index = hash(ip) % len(self.threads)
-        self.threads[thread_index].assign_connection(client_socket, ip)
-    
+        with self.lock:
+            if ip not in self.by_ip:
+                self.by_ip[ip] = self.threads[self.next % len(self.threads)]
+                self.next += 1
+            thread = self.by_ip[ip]
+        thread.assign_connection(client_socket, ip)
+
     def start(self):
         for thread in self.threads:
             thread.start()
@@ -21,66 +24,4 @@ class ThreadPool:
         for thread in self.threads:
             thread.join()
             
-
-class ClientHandlerThread(threading.Thread):
-    """
-    Hilo fijo que maneja conexiones asignadas según addr.
-    """
-    def __init__(self, thread_id, file_lock, barrier):
-        super().__init__()
-        self.thread_id = thread_id
-        self.connections = []
-        self.file_lock = file_lock
-        self.barrier = barrier
-        self.cv = threading.Condition()
-
-    def search_agency_winners(self, id):
-        """Busca los ganadores de una agencia en las apuestas."""
-        winners = []
-        with self.file_lock:
-            for bet in load_bets():
-                if bet.agency == id and has_won(bet):
-                    winners.append(bet.document) 
-        return winners
-    
-    def send_winners(self, client_socket, clientID):
-        logging.info(f"action: sorteo | result: in_progress")
-        winners = self.search_agency_winners(clientID)
-        WinnersAck(winners).send(client_socket)
-        logging.info(f"action: sorteo | result: success")
-    
-    def handle_client(self, client_socket, ip):
-        try:
-            bets, isLastOne, clientID = rcv_bets_in_batch(client_socket)
-
-            # Guardar batch de forma thread-safe
-            with self.file_lock:
-                store_bets(bets)
-                logging.info(f'action: apuesta_recibida | result: success | cantidad: {len(bets)}')
-
-            if isLastOne:
-                faltantes = self.barrier.n_waiting - 1
-                logging.info(f"action: ultima_apuesta_recibida | agencia_id: {clientID} | agency_ip: {ip} | faltantes: {faltantes}")
-                # Último paquete, esperar a todos los clientes
-                self.barrier.wait()
-                self.send_winners(client_socket, clientID)
-                client_socket.close()
-        except OSError as e:
-            logging.error(f"action: apuesta_recibida | result: fail | error: {e}")
-            if client_socket:
-                client_socket.close()
-                
-    def assign_connection(self, client_socket, ip):
-        with self.cv:
-            self.connections.append((client_socket, ip))
-            self.cv.notify()
-
-    def run(self):
-        while True:
-            with self.cv:
-                while not self.connections:
-                    self.cv.wait()  # esperar hasta que haya conexión
-                client_socket, ip = self.connections.pop(0)
-
-            self.handle_client(client_socket, ip)
